Fixed duplicate paragraphs
This removes the duplicate paragraph between CWG and OEWG and attempts to make the CWG paragraph whole again.  However, I do note that this is duplicative of the Working Group page which the language  is not identical. Perhaps there is a better way to restructure this so you do not need to keep this in sync in 2 places.